# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  description: |
    This API allows the user to become part of the WASAText Social Network.
    The user will be able to chat with other users or user groups.
  version: 0.0.1
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      description: "Use the user identifier in place of a Bearer token"
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            description: "Error if Access token is missing or invalid"
            type: object
            properties:
              error:
                description: "Errormessage"
                type: string
                pattern: '^[A-Za-z0-9 ]+$'
                minLength: 1
                maxLength: 255
                example: "Access token is missing or invalid"
    PartnerUsernameNotFound:
      description: The partner username does not exists
      content:
        application/json:
          schema:
            description: "Error if partner username does not exists"
            type: object
            properties:
              error:
                type: string
                pattern: '^[A-Za-z0-9 ]+$'
                minLength: 1
                maxLength: 255
                example: "Partner Username not found"
...
security:
        - bearerAuth: []
...
paths:
  /session:
    post:
      tags:
        - Login
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              description: "Choosen Username"
              type: object
              properties:
                name:
                  type: string
                  pattern: "^.*?$"
                  minLength: 3
                  maxLength: 16
                  example: Maria
                  
        required: true
      security: []
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                description: "Authentification Token"
                type: object
                properties:
                  identifier:
                    # change here if you decide to use an integer
                    # or any other type of identifier
                    type: integer
                    example: 484916884
...
  /groups:
    post:
      tags: 
        - Groups
      summary: adds a user to a group
      operationId: addToGroup
      description: |
        adds a user to an existing group, if the group doesnt exists it creates a new group
      requestBody:
        description: |
          group name and names of the users which should be added
        content:
          application/json:
            schema:
              description: "Requestelement"
              type: object
              properties:
                groupName:
                  description: |
                    Name of the Group to which the users should be added
                  type: string
                  pattern: '^[A-Za-z0-9 ]+$'
                  minLength: 3
                  maxLength: 16
                  example: "WASAGroup"
                  
                names:
                  description: "List of usernames to be added to the group"
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    pattern: '^[A-Za-z0-9 ]+$'
                    minLength: 3
                    maxLength: 16
                    example: "Harry"

                        
        required: true
      responses:
      
        "200":
          description: |
            users successfully added!
          content:
            application/json:
              schema:
                description: "Successmessage"
                type: object
                properties:
                  message:
                    type: string
                    pattern: '^[A-Za-z0-9 ]+$'
                    minLength: 1
                    maxLength: 255
                    example: |
                      Users successfully added to group and group created if didnt existed
        "404":
          $ref: "#/components/responses/PartnerUsernameNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...